requirements.txt
pip
click
click-configfile
click-log
six
future
pytest>=3.1
gitpython
pandas>=0.20.1
Jinja2
jinja2-ospath
cookiecutter
cobra>=0.9.1[sbml]
ruamel.yaml<0.15
travispy
pygithub
travis-encrypt
sympy
sqlalchemy
importlib_resources
numpydoc
pylru
goodtables==1.0.0
